+++ src/api/index.ts
@@ -22,10 +22,10 @@
  * SOFTWARE.
  */
 
-export { getServerInfo } from './getServerInfo'
-export { getColorBars } from './getColorBars'
-export { getDatasets } from './getDatasets'
-export { updateResources } from './updateResources'
-export { getDatasetPlaceGroup } from './getDatasetPlaceGroup'
-export { getTimeSeriesForGeometry } from './getTimeSeries'
-export { HTTPError } from './errors'
+export { getServerInfo } from './getServerInfo';
+export { getColorBars } from './getColorBars';
+export { getDatasets } from './getDatasets';
+export { getDatasetPlaceGroup } from './getDatasetPlaceGroup';
+export { getTimeSeriesForGeometry } from './getTimeSeries';
+export { updateResources } from './updateResources';
+export { HTTPError } from './errors';