+++ src/connected/UserControl.tsx
@@ -23,26 +23,18 @@
  */
 
 import { connect } from 'react-redux';
-import { signIn, signOut } from '../actions/userAuthActions';
 import { AppState } from '../states/appState';
-import { UserAuthStatus } from '../states/userAuthState';
 import UserControl from '../components/UserControl';
+import { updateAccessToken } from '../actions/userAuthActions';
 
 
 // noinspection JSUnusedLocalSymbols
 const mapStateToProps = (state: AppState) => {
-    const userAuthState = state.userAuthState;
-    return {
-        hasAuthClient: !!userAuthState.hasAuthClient,
-        isBusy: userAuthState.status === UserAuthStatus.REQUESTING_SIGN_IN || userAuthState.status === UserAuthStatus.REQUESTING_SIGN_OUT,
-        userInfo: userAuthState.userInfo,
-        accessToken: userAuthState.accessToken,
-    };
+    return {};
 };
 
 const mapDispatchToProps = {
-    signIn,
-    signOut,
+    updateAccessToken
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(UserControl);