+++ CHANGES.md
@@ -3,7 +3,7 @@
 ### Enhancements
 
 * The view app can now display also 2D datasets published by 
-  xcube server (starting with version 0.12).
+  xcube server (starting with xcube version 0.11.3).
 
 ## Changes in version 0.11.1