+++ CHANGES.md
@@ -6,8 +6,14 @@
 
 ### Enhancements
 
-* The view app can now display also 2D datasets published by 
+* A new refresh icon in the main bar now allows updating 
+  server-side resources and refresh the page.
+  For this to work, the configuration setting 
+  `branding.allowRefresh` must be `true`.
+
+* The viewer app can now display also 2D datasets published by 
   xcube server (starting with xcube version 0.11.3).
+
 
 ## Changes in version 0.11.1